fraud_detection_system/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
Flask>=2.0.1
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.6b0
mypy>=0.910
pylint>=2.9.6
requests>=2.26.0
SQLAlchemy>=1.4.23
PyJWT>=2.1.0
python-dateutil>=2.8.2
pytz>=2021.1
tqdm>=4.62.2
joblib>=1.0.1 